+++ src/main/java/com/frejdh/util/job/model/callables/JobOnIdSet.java
@@ -3,6 +3,6 @@ package com.frejdh.util.job.model.callables;
 import com.frejdh.util.job.Job;
 
 @FunctionalInterface
-public interface JobAction {
-	void action(Job jobRef) throws Throwable;
+public interface JobOnIdSet {
+	void onJobIdChange(Job job);
 }