+++ src/main/java/com/frejdh/util/job/persistence/config/DaoPersistenceMode.java
@@ -1,26 +1,26 @@
 package com.frejdh.util.job.persistence.config;
 
-import com.frejdh.util.job.persistence.AbstractDaoServiceImpl;
-import com.frejdh.util.job.persistence.impl.h2.H2DaoService;
-import com.frejdh.util.job.persistence.impl.memory.RuntimeDaoService;
+import com.frejdh.util.job.persistence.AbstractJobQueueDao;
+import com.frejdh.util.job.persistence.impl.h2.H2JobQueueDao;
+import com.frejdh.util.job.persistence.impl.memory.RuntimeJobQueueDao;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
 
 
 public enum DaoPersistenceMode {
-	H2(H2DaoService.class, "h2"),
-	RUNTIME(RuntimeDaoService.class, "runtime", "in-memory", "memory", "internal"),
+	H2(H2JobQueueDao.class, "h2"),
+	RUNTIME(RuntimeJobQueueDao.class, "runtime", "in-memory", "memory", "internal"),
 	CUSTOM(null, "custom");
-	private final Class<? extends AbstractDaoServiceImpl> implementationClass;
+	private final Class<? extends AbstractJobQueueDao> implementationClass;
 	private final List<String> aliases;
 
-	DaoPersistenceMode(Class<? extends AbstractDaoServiceImpl> implementation, String... aliases) {
+	DaoPersistenceMode(Class<? extends AbstractJobQueueDao> implementation, String... aliases) {
 		this.implementationClass = implementation;
 		this.aliases = Arrays.asList(aliases);
 	}
 
-	public Class<? extends AbstractDaoServiceImpl> getImplementationClass() {
+	public Class<? extends AbstractJobQueueDao> getImplementationClass() {
 		return implementationClass;
 	}
 
@@ -38,7 +38,7 @@ public enum DaoPersistenceMode {
 				.findFirst().orElse(null);
 	}
 
-	public static Class<? extends AbstractDaoServiceImpl> toImplementationClass(String arg) {
+	public static Class<? extends AbstractJobQueueDao> toImplementationClass(String arg) {
 		DaoPersistenceMode foundEnum = toEnum(arg);
 		return foundEnum != null ? foundEnum.implementationClass : null;
 	}